- moar coveralls - refactoring for javadoc generation testing
