+++ hello-dangernoodle-module-a/pom.xml.versionsBackup
@@ -6,10 +6,10 @@
   <parent>
     <groupId>io.dangernoodle</groupId>
     <artifactId>hello-dangernoodle</artifactId>
-    <version>0.3.0-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>hello-dangernoodle-module-b</artifactId>
+  <artifactId>hello-dangernoodle-module-a</artifactId>
 
   <build>
     <plugins>
