+++ hello-dangernoodle-module-a/src/integration/java/io/dangernoodle/dummy/a/HelloWorldModuleAIT.java
@@ -1,0 +1,14 @@
+package io.dangernoodle.dummy.a;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+public class HelloWorldModuleAIT
+{
+    @Test
+    public void test()
+    {
+        assertTrue(true);
+    }
+}
