+++ hello-dangernoodle-module-a/src/main/java/io/dangernoodle/dummy/a/HelloWorldModuleA.java
@@ -1,4 +1,4 @@
-package io.dangernoodle.dummy;
+package io.dangernoodle.dummy.a;
 
 /**
  * javadoc for <code>HelloWorldModuleA</code>
@@ -8,10 +8,5 @@ public class HelloWorldModuleA
     public boolean isFoo()
     {
         return true;
-    }
-
-    public static void main(String[] args)
-    {
-        System.out.println("hello world");
     }
 }
