+++ hello-dangernoodle-module-b/pom.xml.versionsBackup
@@ -6,7 +6,7 @@
   <parent>
     <groupId>io.dangernoodle</groupId>
     <artifactId>hello-dangernoodle</artifactId>
-    <version>0.3.0-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
   </parent>
 
   <artifactId>hello-dangernoodle-module-b</artifactId>
