+++ hello-dangernoodle-module-b/src/integration/java/io/dangernoodle/dummy/b/HelloWorldModuleBIT.java
@@ -1,0 +1,14 @@
+package io.dangernoodle.dummy.b;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+public class HelloWorldModuleBIT
+{
+    @Test
+    public void testFoo()
+    {
+        assertTrue(true);
+    }
+}
