+++ hello-dangernoodle-module-b/src/main/java/io/dangernoodle/dummy/b/HelloWorldModuleB.java
@@ -1,10 +1,15 @@
-package io.dangernoodle.dummy;
+package io.dangernoodle.dummy.b;
 
 /**
  * javadoc for <code>HelloWorldModuleA</code>
  */
 public class HelloWorldModuleB
 {
+    public boolean isBaz()
+    {
+        return true;
+    }
+
     public static void main(String[] args)
     {
         System.out.println("hello world");
