+++ hello-dangernoodle-site/pom.xml.versionsBackup
@@ -7,7 +7,7 @@
   <parent>
     <groupId>io.dangernoodle</groupId>
     <artifactId>hello-dangernoodle</artifactId>
-    <version>0.3.0-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
   </parent>
 
   <name>${project.artifactId}</name>
@@ -49,15 +49,6 @@
       <plugin>
         <groupId>com.github.hazendaz.maven</groupId>
         <artifactId>coveralls-maven-plugin</artifactId>
-        <configuration>
-          <jacocoReports>
-            <jacocoReport>${project.reporting.outputDirectory}/jacoco-aggregate/jacoco.xml </jacocoReport>
-          </jacocoReports>
-          <sourceDirectories>
-            <directory>../hello-dangernoodle-module-a/src/main/java</directory>
-            <directory>../hello-dangernoodle-module-b/src/main/java</directory>
-          </sourceDirectories>
-        </configuration>
       </plugin>
     </plugins>
   </build>
