+++ pom.xml.versionsBackup
@@ -10,7 +10,7 @@
   </parent>
 
   <artifactId>hello-dangernoodle</artifactId>
-  <version>0.3.0-SNAPSHOT</version>
+  <version>0.2.1-SNAPSHOT</version>
   <packaging>pom</packaging>
 
   <name>${project.artifactId}</name>
